Refactor to keep files better organized
